requirements.txt
flask
python-dotenv
beautifulsoup4
requests
langchain
langchain-community
langchain-core
sentence-transformers
chromadb
openai
langchain-openai 